server/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
redis==5.0.4
rq==1.16.2
fakeredis==2.23.2
python-jose[cryptography]==3.3.0
numpy>=2.0.0
Pillow>=10.3.0
zstandard>=0.23.0
psutil==5.9.8
msgpack==1.0.8
httpx==0.27.0
pydantic==2.7.1
